+++ src/functions/web/handleWSMessage.ts
@@ -7,6 +7,8 @@ import { getPreviousHours } from "./handleHTTPRequest"
 export default function handleWSConnect(ws: WebSocket<WebSocketContext>, message: ArrayBuffer, ctg: GlobalContext) {
 	const { custom, ctx } = ws.getUserData()
 
+	ctg.logger.debug('WebSocket message with bytelen', message.byteLength, 'recieved')
+
 	ctx.response.content = Buffer.allocUnsafe(0)
 	ctx.previousHours = getPreviousHours()
 	ctx.body.raw = Buffer.from(message)